Add setup for preprocessing with redis
